9-3 daily update.
[FLE]Added ceramic pot and hearth.
[Nebula]Prepare to remove some unused interfaces.
[FLE]Added adobe and crushed bones.
[FLE]The thermal energy net will has a big changed.

+++ src/main/java/nebula/common/tile/ITilePropertiesAndBehavior.java
@@ -270,7 +270,7 @@ public interface ITilePropertiesAndBehavior
 		
 		boolean onBurningTick(Random rand, Direction fireSourceDir, IBlockState fireState);
 		
-		void onHeatChange(Direction direction, double amount);
+		void onHeatChange(Direction direction, long amount);
 	}
 	
 	public static interface ITB_AddDestroyEffects
